* 3.7.0 released.   (tag "rel-3_7_0") * configure.in, mlterm.spec, main/version.h.in,   MLTermPty.c, android/jni/version.h:   updated for 3.7.0. * etc/main, etc/main.ja: Updated. * fb/x_color.c, fb/x_display.c: The type of the first argument of   x_cmap_get_closest_color() is reverted from u_int to u_long.

+++ xwindow/fb/x_display.c
@@ -2510,7 +2510,7 @@ x_display_reset_input_method_window(void)
 /* seek the closest color */
 int
 x_cmap_get_closest_color(
-	u_int *  closest ,
+	u_long *  closest ,
 	int  red ,
 	int  green ,
 	int  blue
